// src/lib/dto/asset.dto.ts
export interface AssetDto {
  id: string;
  number: string;
  description: string;
  description2: string | null;
  type: string | null;
  assetType: string | null;
  status: string | null;
  propertyNo: string | null;
  propertyDescription: string | null;
  address: string | null;
  address2: string | null;
  city: string | null;
  postCode: string | null;
  county: string | null;
  countryRegionCode: string | null;
  streetName: string | null;
  streetNumber: string | null;
  floor: string | null;
  composedAddress: string | null;
  googleUrl: string | null;
  yearOfConstruction: number | null;
  builtAreaM2: number | null;
  commercialDescription: string | null;
  cadastralReference: string | null;
  cadastralUrl: string | null;
  cadastralAssetValue: number | null;
  cadastralConstructionValue: number | null;
  totalCadastralAssetValue: number | null;
  totalCadastralConstructionValue: number | null;
  ownerName: string | null;
  lastRentalPrice: number | null;
  minimumRentalPrice: number | null;
  lastContractPrice: number | null;
  referencePriceMin: number | null;
  referencePriceMax: number | null;
  salesPrice: number | null;
  minimumSalesPrice: number | null;
  managed: boolean;
  acquired: boolean;
  blocked: boolean;
  underMaintenance: boolean;
  insured: boolean;
  hasComments: boolean;
  image: string | null;
  lastDateModified: string | null;
}